+++ src/wallet/rpcwallet.h
@@ -5,6 +5,7 @@
 #ifndef BITCOIN_WALLET_RPCWALLET_H
 #define BITCOIN_WALLET_RPCWALLET_H
 
+#include <map>
 #include <string>
 
 class CCoinControl;
@@ -13,6 +14,8 @@ class CWallet;
 class CWalletTx;
 class JSONRPCRequest;
 class UniValue;
+
+typedef std::map<std::string, std::string> mapValue_t;
 
 void RegisterWalletRPCCommands(CRPCTable &t);
 
@@ -27,10 +30,11 @@ CWallet *GetWalletForJSONRPCRequest(const JSONRPCRequest& request);
 std::string HelpRequiringPassphrase(CWallet *);
 void EnsureWalletIsUnlocked(CWallet *);
 bool EnsureWalletIsAvailable(CWallet *, bool avoidException);
-void SendMoneyToScript(CWallet* pwallet, const CScript& scriptPubKey,
-                       const CTxIn* withInput, CAmount nValue,
-                       bool fSubtractFeeFromAmount, CWalletTx& wtxNew,
-                       const CCoinControl& coin_control);
+CTransactionRef SendMoneyToScript(CWallet* pwallet, const CScript& scriptPubKey,
+                                  const CTxIn* withInput, CAmount nValue,
+                                  bool fSubtractFeeFromAmount,
+                                  const CCoinControl& coin_control,
+                                  mapValue_t mapValue, std::string fromAccount);
 
 UniValue getaddressinfo(const JSONRPCRequest& request);
 UniValue signrawtransactionwithwallet(const JSONRPCRequest& request);